backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
requests==2.31.0
python-dotenv==1.0.1
elasticsearch==8.13.0
pdfkit==1.0.0
jinja2==3.1.3
